refactor: list page 폴더 네이밍 수정, route 상수 수정
