+++ src/route.js
@@ -1,9 +1,9 @@
 export const ROUTE = {
   home: { pageTitle: '홈 화면', route: '/' },
-  cardList: { pageTitle: '보유 카드', route: '/list' },
-  addCard: { pageTitle: '카드 추가', route: '/list/add' },
-  editCard: { pageTitle: '카드 수정', route: '/list/edit' },
-  cardSuccess: { pageTitle: '별칭 추가', route: '/list/success' },
+  cardList: { pageTitle: '보유 카드', route: '/cardlist' },
+  addCard: { pageTitle: '카드 추가', route: '/cardlist/add' },
+  editCard: { pageTitle: '카드 수정', route: '/cardlist/edit' },
+  cardSuccess: { pageTitle: '별칭 추가', route: '/cardlist/success' },
 };
 
 export const PAGE_TITLE = {
